pyproject.toml
[project]
name = "jpgconverter"
version = "1.0.0"
description = "JPEG 批量转换器 - 支持 HEIC/AVIF/JXL 双向转换"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=10.0.0",
    "pillow-heif>=0.13.0",
    "pillow-avif-plugin>=1.5.5",
    "pillow-jxl-plugin>=1.3.0",
]

[project.scripts]
jpgconverter = "jpgconverter.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jpgconverter"]
